Feat: 모달 표기 위치 변경 (MM-147)
