+++ apps/react/src/app/page.tsx
@@ -3,6 +3,7 @@ import { createFileRoute, useNavigate } from '@tanstack/react-router'
 
 import HeartIcon from '@/assets/icons/heart.svg'
 import malmoLogo from '@/assets/images/malmo-logo-small.png'
+import { AnniversaryEditSheet } from '@/features/anniversary'
 import { getAttachmentType } from '@/features/attachment'
 import { AttachmentTestBanner } from '@/features/attachment/ui/attachment-test-banner'
 import { AttachmentTypeCards } from '@/features/attachment/ui/attachment-type-cards'
@@ -11,6 +12,7 @@ import { useChatRoomStatusQuery } from '@/features/chat/hooks/use-chat-queries'
 import { ChatEntryCard } from '@/features/chat/ui/chat-entry-card'
 import { usePartnerInfo } from '@/features/member'
 import { useAppNotifications } from '@/features/notification'
+import { useProfileEdit } from '@/features/profile'
 import { TodayQuestionSection, useTodayQuestion } from '@/features/question'
 import { wrapWithTracking } from '@/shared/analytics'
 import { BUTTON_NAMES, CATEGORIES } from '@/shared/analytics/constants'
@@ -42,6 +44,7 @@ function HomePage() {
 
   const { data: todayQuestion } = useTodayQuestion()
   const { data: partnerInfo } = usePartnerInfo()
+  const profileEdit = useProfileEdit()
 
   // D-day 계산
   const dDay = calculateDDay(userInfo.startLoveDate)
@@ -84,6 +87,11 @@ function HomePage() {
     }
   })
 
+  // 기념일 시트 열기 핸들러
+  const handleAniversaryEdit = wrapWithTracking(BUTTON_NAMES.OPEN_ANNIVERSARY_SHEET, CATEGORIES.PROFILE, () =>
+    profileEdit.openAnniversarySheet()
+  )
+
   return (
     <div className="has-bottom-nav flex h-full flex-col bg-white pt-[60px]">
       {/* 헤더 */}
@@ -92,7 +100,10 @@ function HomePage() {
 
         {/* D-day */}
         {isPartnerConnected && (
-          <div className="flex h-8 items-center rounded-[30px] border border-gray-iron-200 px-4 py-[5px]">
+          <div
+            className="flex h-8 items-center rounded-[30px] border border-gray-iron-200 px-4 py-[5px]"
+            onClick={handleAniversaryEdit}
+          >
             <HeartIcon className="h-4 w-4" />
             <span className="body2-semibold ml-[9px] text-gray-iron-950">D+{dDay}</span>
           </div>
@@ -124,6 +135,12 @@ function HomePage() {
 
       {/* 하단 네비게이션 */}
       <BottomNavigation />
+
+      <AnniversaryEditSheet
+        isOpen={profileEdit.isAnniversarySheetOpen}
+        onOpenChange={profileEdit.setAnniversarySheetOpen}
+        onSave={wrapWithTracking(BUTTON_NAMES.SAVE_ANNIVERSARY, CATEGORIES.PROFILE)}
+      />
     </div>
   )
 }
